test: achieve 100% coverage
